@@ -1,51 +1,45 @@
 {
-  "name": "sidur",
+  "name": "Sidur",
   "version": "0.1.0",
-  "private": false,
+  "private": true,
+  "homepage": "https://chenpeleg.github.io/Sidur",
   "homepage-gh-pages": "https://chenpeleg.github.io/Sidur",
-  "dependencies": {
-    "@emotion/react": "^11.4.1",
-    "@emotion/styled": "^11.3.0",
-    "@mui/icons-material": "^5.0.3",
-    "@mui/material": "^5.0.3",
-    "@mui/system": "^5.0.3",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/enzyme": "^3.10.10",
-    "@types/jest": "^26.0.24",
-    "@types/node": "^12.20.27",
-    "@types/react": "^17.0.26",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-router-dom": "^5.3.3",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.6",
-    "final-form": "^4.20.4",
-    "gh-pages": "^4.0.0",
-    "node-sass": "^6.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-final-form": "^6.5.7",
-    "react-redux": "^7.2.5",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "4.0.3",
-    "redux": "^4.1.1",
-    "redux-devtools-extension": "^2.13.9",
-    "typescript": "^4.4.3",
-    "web-vitals": "^1.1.2"
+  "engines": {
+    "node": "18"
   },
-  "jest": {
-    "watchPathIgnorePatterns": [
-      "<rootDir>/src/__tests__"
-    ]
+  "dependencies": {
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@mui/icons-material": "^5.10.9",
+    "@mui/lab": "^5.0.0-alpha.106",
+    "@mui/material": "^5.10.12",
+    "@mui/x-date-pickers": "^5.0.6",
+    "@redux-devtools/extension": "^3.2.3",
+    "@types/node": "^16.11.56",
+    "@types/react": "^18.0.18",
+    "@types/react-dom": "^18.0.6",
+    "date-fns": "^2.29.3",
+    "final-form": "^4.20.7",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-final-form": "^6.5.9",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^6.4.3",
+    "react-scripts": "5.0.1",
+    "redux": "^4.2.0",
+    "typescript": "^4.8.2",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test --silent=false",
+    "test": "react-scripts test --silent=false --watchAll=false",
+    "test:last": "react-scripts test -t last --silent=false --watchAll=false",
+    "test:watch": "react-scripts test",
     "eject": "react-scripts eject",
     "predeploy": "react-scripts build",
-    "deploy": "gh-pages -d build"
+    "deploy": "gh-pages -d build",
+    "format": "prettier --write ."
   },
   "eslintConfig": {
     "extends": [
@@ -66,8 +60,16 @@
     ]
   },
   "devDependencies": {
-    "@types/enzyme-adapter-react-16": "^1.0.6",
-    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.5",
-    "redux-devtools": "^3.7.0"
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/jest": "^29.2.1",
+    "@typescript-eslint/eslint-plugin": "^5.42.0",
+    "@typescript-eslint/parser": "^5.42.0",
+    "eslint": "^8.26.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "gh-pages": "^4.0.0",
+    "prettier": "^2.7.1"
   }
 }
