@@ -1,16 +1,20 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './hoc/App';
-import reportWebVitals from './reportWebVitals';
-import {Root} from './hoc/Root';
+import React from "react";
+import "./index.css";
+import App from "./hoc/App";
+import reportWebVitals from "./reportWebVitals";
+import { Root } from "./hoc/Root";
+import ReactDOM from "react-dom/client";
 
-
-ReactDOM.render(
+const root = ReactDOM.createRoot(
+  document.getElementById("root") as HTMLElement
+);
+root.render(
+  <React.StrictMode>
     <Root>
-        <App/>
-    </Root>,
-    document.getElementById('root')
+      <App />
+    </Root>
+    ,
+  </React.StrictMode>
 );
 
 reportWebVitals();
