@@ -2,10 +2,7 @@
 // allows you to do things like:
 // expect(element).toHaveTextContent(/react/i)
 // learn more: https://github.com/testing-library/jest-dom
-import '@testing-library/jest-dom';
-import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
-import {configure} from 'enzyme';
-import React from 'react'
+import "@testing-library/jest-dom";
+import React from "react";
 
-React.useLayoutEffect = React.useEffect
-configure({adapter: new Adapter()})
+React.useLayoutEffect = React.useEffect;
